Use MARATHON_SERVICE_SCHEME to define the scheme of the service link
The service link under the app view and task view uses a scheme provided
as a value of the label called MARATHON_SERVICE_SCHEME. The value could be
either `http` or `https`.

One can also define the scheme by port by setting a
MARATHON_SERVICE_SCHEME_<N> where N is the index of the port for which to
use the scheme. The value must be either `http` or `https`.

If no label is provided, the link scheme is supposed to be http by
default.