Containerizer type switch when using application edit form.
The containerizer type is switching from MESOS to DOCKER when
any change is done in the application edit form. It's due to the fact
that the field is auto-detected by the presence of the key 'docker' in the
app model. However the presence of this key is not sufficient to correctly
detect the containerizer in use since Mesos containerizer can also use Docker
images (so also declaring 'docker' key).

I also took the opportunity to add the field in the form so that the user can
switch without going through the UI.